turn creatId value into long type

+++ localfullnode2/src/main/java/one/inve/localfullnode2/gossip/GossipDependency.java
@@ -74,7 +74,7 @@ public class GossipDependency implements GossipDependent, DependentItemConcerned
 	}
 
 	@Override
-	public int getCreatorId() {
+	public long getCreatorId() {
 		return creatorId.get();
 	}
